Rules component refactored to dialog. Created SavedRule component. Work on UX for RulesView
