+++ bloxiom/public/index.html
@@ -8,7 +8,12 @@
       name="description"
       content="Web site created using create-react-app"
     />
-    <title>React App</title>
+    <title>Bloxiom</title>
+    <style type="text/css">
+      body {
+        margin: 0;
+      }
+    </style>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
